added List class and shared_formula cell management
